refactored crud components and combined them on single pages
