+++ frontend/src/utilities/formatter.ts
@@ -4,5 +4,6 @@ const formatter = new Intl.NumberFormat('en-US', {
 });
 
 export const formatPrice = (number: number) => {
+  if (isNaN(number)) return number;
   return formatter.format(number);
 };
